AudioFlinger: re-organize position adjustment for restarted DirectOutputs
Move logic for resetting frame position for direct outputs post standby.
Output can be put to standby when track is stopped/idle as well as
while still in active state where linear pcm formats is used and track
can be put to standby and restored back without affecting playback.
AudioStreamOut layer is not able to distinguish between standby on
active/stopped track and hence was doing postion adjustment even when
it was not required.
Moving the logic to DirectOutput thread so that it can check track state
and do position adjustment only when standby was triggered due to track
being stopped/idle.

Change-Id: I8405e9f9a96ee8aef3281e8ed45506b60d5a5b44
Signed-off-by: Marco Zanin (B--B) <mrczn.bb@gmail.com>